Enemy AI updated
Added a list of Active Enemies in the scene on the GameManager for a reference for Enemy AI so it Enemies can push away from each other and not stack up. Added Enemy AI script with direction and distance checking to chase player and not stack up on other enemies.