Experience Updated
Added a list of experience brackets for each level up. Updated the UI to display with experience lists and check for level up. Still need to add a full list of level ups or a continuous way to change the experience needed to level up. Enemies also can be assigned a Experience worth amount.