Health Bar Bugged still
Still attempting to fix health bar bug. The function that changes the health bar UI is not being called, the fill value on the slider is being changed outside of that script. Not sure if the type of value being set is causing the problem. Seems to only happen when the health is low. Hard to replicate currently.