Health Drop and IInteract
Updated Experience drop to have Interact Interface so when the player interacts with anything with the IInteract Interface it will call the same function.  Impelmented Health Drops with a new prefab and enemies will now spawn them instead of exp for now. Next need a way to determine when to drop an exp or a health drop.