Enemy Spawner Phases
Added Enemy spawn phases. In the inspector we add a phase by setting a time when the phase starts, what cooldown spawn timer should be and what enemies are spawned in the phase. Currently just have it on a timer to change phases but can later set it to player level or the game time clock if needed.