Updated EnemySpawner
Updated Enemyspawner with player level up changing the phase instead of the game timer. Idea to add config to the phases to only change cooldown or add some new enemies rather than change a whole phase. Maybe phases should be seperated from the enemy Spawner to use in other parts of the game like UI or sound later on.