Refactor input/ui upgrade panel
Refactored a few scripts for future abilities. Input was moved to the Player script away from the movement. Player Script will handle all inputs the player makes from now on. Moved functionality of upgrade panel to the ability controller since the logic was more for controlling abilities and a way to keep track of input abilities for the player script.