Treasure Chest/Follow Arrow UI
Added a treasure chest sprite in the game, then added a UI arrow image that will point towards the object getting the distance between the player and the target object. Need to add functionality of treasure, maybe require certain level, and then some way to open it. Also add a way to change target of the arrow.