End Game Panel/Main Menu Scene
Created a temp Main menu scene to load when the main menu button is clicked on the end game panel. Added the end game panel to pop up when the game manager script calls for the end of the game. Still need to clean up the panel and add functionality to the exit button. Nothing in the main menu scene yet.